requirements.txt
PixivPy
attrs
